lag witch broadcast

+++ MinerClientProgramm/MinerStatus.cs
@@ -11,7 +11,7 @@ namespace MinerClientProgramm
         public string user { get; set; }
         public string pool { get; set; }
 
-        public int GPUtemp { get; set; }
+        public double GPUtemp { get; set; }
 
         public bool running { get; set; }
     }
